requirements.txt
Django==1.9
psycopg2==2.6.1
wheel==0.24.0
django-model-utils==2.4
django-admin-bootstrapped==2.5.6
django-bootstrap3==6.2.2
django-braces